DCPU16.Assembler

Description

Assembler.

Works with DCPU16.Instructions. Does no optimizations, just converts labels to direct addresses and produces machine code.

Needs to be extended to support relative addressing. Extending labels to do arithmetic: sub pc,this-loop is a necessary step. I see this as more of an optimization problem, since these kinds of rewrites would be neat to perform automatically...

We'll see which direction it gets approached from.

Documentation

assemble :: Vector Instruction -> ByteString

Turn labels into direct addresses, and output machine code.

Utility

labelAddrs :: Vector Instruction -> Map ByteString Word16

Build address lookup table for label definitions.