DCPU16.Assembler.Parser

Description

Parser for what seems to be DCPU16 assembly.

There is some ambiguity between sources: the specification uses uppercase a lot (which I'd rather put in as an option later, with the strict implementation being default).

A screenshot also shows indirect mode being done with [] instead of (). Go figure.

Documentation

parseFile :: String -> IO (Maybe (Vector Instruction))