DCPU16.Disassembler

Description

Disassembler.

I should really convert these to use Vector Word16, rather than ByteString.

It won't be useful for command line tools, but make more sense for everything else.

Documentation

disassemble :: ByteString -> Vector Instruction

Disassembler.

Disassembly can fail at the very end of a buffer, if a partial multi-word instruction is encountered. I'll eventually treat those as data words instead, but for now it crashes... And is a point of weakness to keep in mind for interpreters. What will the official one do?