DCPU16.Instructions

Description

Complete abstract description of the DCPU-16 instruction set.

Based on Version 1.1 of the DCPU-16 Specification by Mojang, retrieved from 0x10c.com.

Contains a trivial Label extension, which isn't present in machine code but is useful for dealing with assembly.

Documentation

data Instruction

Abstract DCPU-16 instruction set.

Can be read and written as machine code via the Serialize instance.

Constructors

Basic BasicOp Operand Operand

NonBasic NonBasicOp Operand

Data Word

Label ByteString
Not present in machine code, for assembler utility only.

Comment Bool String
Boolean true if comment is alone on its own line.

Operands

data Operand

Values instructions operate on.

Direct operands use the value passed to them.

Indirect operands treat that value as an address for a specific word in memory.

Constructors

Direct Register

Indirect Register
At address [register].

Offset Word Register
At address [next word + register].

Pop

Peek

Push

SP

PC

O
Overflow.

IndirectLiteral Word

DirectLiteral Word

ShortLiteral Word
Restricted to 0x00-0x1f, 5 bits.

data Register

Constructors

A

B

C

X

Y

Z

I

J

data Word

Constant data.

Assembly may use adresses of labels to initialize such data: since the address may not be known immediately, the label extension is added.

Constructors

Const Word16

LabelAddr ByteString

Operations

data BasicOp

Constructors

SET

ADD

SUB

MUL

DIV

MOD

SHL

SHR

AND

BOR

XOR

IFE

IFN

IFG

IFB

data NonBasicOp

Constructors

JSR

Reserved Word16
Opcode not defined yet.